requirements.txt
# Third-party modules required for Ragbot.AI
# Updated: December 2025 for compatibility
anthropic>=0.31.0
appdirs
babel
boto3
google-generativeai>=0.7.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
langchain-openai>=0.3.0
litellm>=1.78.0
openai>=2.0.0
pathlib
pytest
pytest-mock
python-dotenv
pyyaml
streamlit
tiktoken
watchdog

# RAG dependencies
qdrant-client>=1.7.0
sentence-transformers>=2.2.0

# FastAPI dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=1.6.0
pydantic>=2.0.0
